% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtrans.R
\name{Gtrans}
\alias{Gtrans}
\title{G-trasformation function}
\usage{
Gtrans(data)
}
\arguments{
\item{data}{The dataset to be transforms. The dataset can be discretein all columns, continuous in all columns or a mixture of continuous columns and discrete columns.}
}
\value{
A transformed continuous dataset with the same copula as the input dataset and margianl normal(0,1).
}
\description{
function to transform discrete or mixture of discrete and continuous datasets to continuous datasets with marginal normal(0,1).
}
