% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{overlap_mat_sim}
\alias{overlap_mat_sim}
\title{Overlap matrices for simulated data}
\format{
A list of 3 matrices, each 5x5, representing overlap between classes
}
\usage{
overlap_mat_sim
}
\description{
This is a list containing three overlap matrices corresponding to the
sim_data datasets, showing class separability.
}
\keyword{datasets}