% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sarscov2.us.variants}
\alias{sarscov2.us.variants}
\title{COVID-19 US variants dataset}
\format{
A data frame of 140 observations and 14 variables.
\describe{
    \item{group}{The date.}
    \item{type}{weighted}
    \item{region}{Region of the US labelled by numbers.}
    \item{B.1.1.194, B.1.1.7, B.1.351, B.1.525, B.1.526, B.1.621, B.1.628, B.1.637, Delta, Other*, P.1}{COVID-19 variants compositions.}
}
}
\usage{
sarscov2.us.variants
}
\description{
This is a compositional dataset of the COVID-19 variants in the US from 6/19/2021 to 9/18/2021.
}
\keyword{datasets}
