% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Simulated datasets for testing}
\format{
A list of 3 data frames, each with 500 observations and 6 variables:
\describe{
    \item{class}{Factor with 5 levels representing different classes}
    \item{X1, X2, X3, X4, X5}{Numeric variables with simulated data}
}
}
\usage{
sim_data
}
\description{
This is a list containing three simulated datasets, each with 500 observations
and 5 classes, used for testing visualization methods.
}
\keyword{datasets}