% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wine}
\alias{wine}
\title{Chemical compositions of wine}
\format{
A data frame of 178 observations and 14 variables:
    \describe{
        \item{cultivar}{The cultivar where the wine is produced}
        \item{other variables}{The content of chemical compositions of the wine}
    }
}
\usage{
wine
}
\description{
The dataset contains chemical compositions of wines from 3 cultivars
}
\keyword{datasets}
