% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtrans.R
\name{Gtrans}
\alias{Gtrans}
\title{G-trasformation function}
\usage{
Gtrans(data, cl = NULL, VariableSelection = FALSE, p_threshold = 0.05, ...)
}
\arguments{
\item{data}{The dataset to be transforms. The dataset can be discretein all columns, continuous in all columns or a mixture of continuous columns and discrete columns.}

\item{cl}{The class information of the dataset. This is not required when \code{VariableSelection = FALSE}.}

\item{VariableSelection}{Logical. If true, anova will be performed to each variable to see whether there is a difference among groups for that variable. The varaible associated with Bonferroni adjusted p-value larger than a threshold will be removed.}

\item{p_threshold}{The threshold for adjusted p-value in variable selection when \code{VariableSelection = TRUE}.}
}
\value{
A transformed continuous dataset with the same copula as the input dataset and margianl normal(0,1).
}
\description{
function to transform discrete or mixture of discrete and continuous datasets to continuous datasets with marginal normal(0,1).
}
