% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radviz3d.R
\name{radialvis3d}
\alias{radialvis3d}
\title{3D Radial Visualization function}
\usage{
radialvis3d(data, cl = NULL, color = NULL, axis = FALSE,
  coord.labels = colnames(data), coord.font = 2, coord.cex = 1.1,
  class.labels = levels(factor(cl)), class.labels.locations = NULL,
  opt.anchor.order = FALSE, ...)
}
\description{
3D Radial Visualization function
}
